feat: add the macro for some extension field(Fp6, Fq6) tests
