+++ src/tests/field.rs
@@ -590,6 +590,94 @@ macro_rules! field_testing_suite {
         }
     };
 
+    ($ext_field: ident, "f6_tests", $base_field: ident) => {
+        #[test]
+        fn test_f6_mul_nonresidue() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            let nqr = $ext_field {
+                c0: $base_field::zero(),
+                c1: $base_field::one(),
+                c2: $base_field::zero(),
+            };
+
+            for _ in 0..1000 {
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+                a.mul_by_nonresidue();
+                b.mul_assign(&nqr);
+
+                assert_eq!(a, b);
+            }
+        }
+
+        #[test]
+        fn test_f6_mul_by_1() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let c1 = $base_field::random(&mut rng);
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+
+                a.mul_by_1(&c1);
+                b.mul_assign(&$ext_field {
+                    c0: $base_field::zero(),
+                    c1,
+                    c2: $base_field::zero(),
+                });
+
+                assert_eq!(a, b);
+            }
+        }
+
+        #[test]
+        fn test_f6_mul_by_01() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let c0 = $base_field::random(&mut rng);
+                let c1 = $base_field::random(&mut rng);
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+
+                a.mul_by_01(&c0, &c1);
+                b.mul_assign(&$ext_field {
+                    c0,
+                    c1,
+                    c2: $base_field::zero(),
+                });
+
+                assert_eq!(a, b);
+            }
+        }
+
+        #[test]
+        fn test_squaring() {
+            let mut rng = XorShiftRng::from_seed([
+                0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06, 0xbc,
+                0xe5,
+            ]);
+
+            for _ in 0..1000 {
+                let mut a = $ext_field::random(&mut rng);
+                let mut b = a;
+                b.mul_assign(&a);
+                a.square_assign();
+                assert_eq!(a, b);
+            }
+        }
+    };
+
     ($ext_field: ident, "frobenius", $frobenius_param: expr) => {
         #[test]
         fn test_frobenius() {
